backend/django_code/pyproject.toml
[project]
name = "django_code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "datetime>=6.0",
    "django>=6.0",
    "django-debug-toolbar>=6.1.0",
    "pywatchman>=3.0.0",
    "djangorestframework>=3.15.0",
    "django-cors-headers>=4.3.0",
    "psycopg2-binary>=2.9.0",
    "djangorestframework-stubs>=3.16.6",
    "pillow>=12.0.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
]
prod = [
    "uvicorn[standard]>=0.28.0",
]
